Convert all projects to proper sdk style and multi target
