+++ .gitignore
@@ -1,6 +1,9 @@
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
 
+## Android Auto-generated files
+Resource.designer.cs
+
 # User-specific files
 *.suo
 *.user
@@ -27,6 +30,7 @@ bld/
 
 # MSTest test Results
 [Tt]est[Rr]esult*/
+[Tt]est[Rr]esult*
 [Bb]uild[Ll]og.*
 
 # NUNIT
@@ -37,6 +41,12 @@ TestResult.xml
 [Dd]ebugPS/
 [Rr]eleasePS/
 dlldata.c
+
+# .NET Core
+project.lock.json
+project.fragment.lock.json
+artifacts/
+**/Properties/launchSettings.json
 
 *_i.c
 *_p.c
@@ -62,6 +72,10 @@ dlldata.c
 *.pidb
 *.svclog
 *.scc
+*.exe
+*.bak
+*.cache
+*.lib
 
 # Chutzpah Test files
 _Chutzpah*
@@ -102,6 +116,7 @@ _TeamCity*
 # NCrunch
 _NCrunch_*
 .*crunch*.local.xml
+*.ncrunchproject
 
 # MightyMoose
 *.mm.*
@@ -136,15 +151,19 @@ publish/
 
 # NuGet Packages
 *.nupkg
-# Nuget Artifacts
-*.nuget.props
-*.nuget.targets
 # The packages folder can be ignored because of Package Restore
 **/packages/*
 # except build/, which is used as an MSBuild target.
 !**/packages/build/
 # Uncomment if necessary however generally it will be regenerated when needed
 #!**/packages/repositories.config
+# NuGet v3's project.json files produces more ignoreable files
+*.nuget.props
+*.nuget.targets
+
+#Allow NuGet.exe (do not ignore)
+!NuGet.exe
+*.orig
 
 # Windows Azure Build Output
 csx/
@@ -198,15 +217,46 @@ FakesAssemblies/
 # Visual Studio 6 workspace options file
 *.opt
 
-# generated code 
-Resource.Designer.cs
+#ignore thumbnails created by windows
+Thumbs.db
 
-# fake
-.fake/
-FAKE.**.**.**/
+# Xcode
+.DS_Store
+*/build/*
+*.pbxuser
+!default.pbxuser
+*.mode1v3
+!default.mode1v3
+*.mode2v3
+!default.mode2v3
+*.perspectivev3
+!default.perspectivev3
+xcuserdata
+profile
+*.moved-aside
+DerivedData
+.idea/
+*.hmap
+*.xccheckout
 
-# Build
-out/
-*.lock.json
-src/samsung/sdk/android.jar
-*.bak
+#CocoaPods
+Pods
+nuspec/nuget.exe
+
+# NuGet
+packages/
+repositories.config
+
+#Build
+project.lock.json
+tools/*
+!tools/packages.config
+artifacts/
+*.nupkg
+*.binlog
+
+#linting
+.sonarqube/
+
+# dotnet tools
+.store/
