+++ appveyor.yml
@@ -1,0 +1,45 @@
+-
+  branches:
+      only:
+        - master
+        - /release\/.*/
+  version: 1.0.{build}
+  image: Visual Studio 2017
+  before_build:
+  - cmd: dotnet --info
+  - cmd: dotnet tool install --tool-path . SignClient
+  environment:
+    NUGET_SOURCE: https://www.nuget.org/api/v2/package
+    NUGET_APIKEY:
+      secure: 
+    SIGNING_USER:
+      secure: 
+    SIGNING_SECRET:
+      secure: 
+  build_script:
+  - ps: .\build.ps1
+  test: off
+  cache:
+  - '%USERPROFILE%\.nuget\packages -> **\project.json'  # project.json cache
+-
+  branches:
+      only:
+        - develop
+  version: 1.0.{build}
+  image: Visual Studio 2017
+  before_build:
+  - cmd: dotnet --info
+  - cmd: dotnet tool install --tool-path . SignClient
+  environment:
+    NUGET_SOURCE: https://www.myget.org/F/mvvmcross/api/v2/package
+    NUGET_APIKEY:
+      secure: 
+    SIGNING_USER:
+      secure: 
+    SIGNING_SECRET:
+      secure: 
+  build_script:
+  - ps: .\build.ps1
+  test: off
+  cache:
+  - '%USERPROFILE%\.nuget\packages -> **\project.json'  # project.json cache
