--- src/.build/MvvmCross.Plugins.Fingerprint.nuspec
@@ -1,84 +1,0 @@
-<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-  <metadata>
-    <id>MvvmCross.Plugins.Fingerprint</id>
-    <version>1.4.7</version>
-    <title>MvvmCross Fingerprint Plugin</title>
-    <authors>Sven-Michael Stübe</authors>
-    <owners>Sven-Michael Stübe</owners>
-    <licenseUrl>https://github.com/smstuebe/xamarin-fingerprint/blob/master/LICENSE</licenseUrl>
-    <projectUrl>https://github.com/smstuebe/xamarin-fingerprint</projectUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Authenticate a user via fingerprint from a cross platform API.</description>
-    <summary>Authenticate a user via fingerprint from a cross platform API.</summary>
-    <tags>xamarin mvvmcross mvvm monodroid Xamarin.iOS10 ios fingerprint authenticate Xamarin.Android android touchid biometric windows uwp</tags>
-    <iconUrl>https://raw.githubusercontent.com/smstuebe/xamarin-fingerprint/master/doc/xamarin_fingerprint.png</iconUrl>
-    <releaseNotes>
-      [1.4.7]
-      - #101 FingerprintAuthenticationResultStatus and FingerprintAvailability now can express that the user has denied the access to the authentication method
-      [1.4.6]
-      - CrossFingerprint is now setable for mocking during unit tests (thanks @ArtjomP)
-      - #51 allows setting the `DefaultColor` to change the color of the fingerprint icon
-      - #86 android fallback button is now hidden if AllowAlternativeAuthentication is false (thanks @fedemkr)
-      - #91 new API `GetAuthenticationTypeAsync` to retrieve the biometric auth type
-      - #84 missing resource ids should be available again
-      - #83 prevent NullRefereneceException and catch others if service is not available on Android
-      - #75 fix crash at onPause on Samsung devices
-      - #73 display help string for recoverable errors
-      - #29 fix crash on device lock during authentication
-      [1.4.5]
-      - #53, #70 fixed possible crash on some samsung devices without fingerprint sensor
-      [1.4.4]
-      - #60 allow alternative authentication via PIN / password (iOS only)
-      - allow custon fallback title on Mac
-      [1.4.3]
-      - #45: fixed UWP nuget package
-      [1.4.2]
-      - #43: fixed crash when fast tapping on cancel or fallback on Android
-      [1.4.1]
-      - #40: fixed crash on iOS &lt; 8.0
-      [1.3.0]
-      - support for Samsung devices
-      - support for .NET standard
-      - support for dialog localization
-      - animations on Android
-      - bugfixes
-      [1.2.0]
-      - fixed plugin loader bug
-      - iOS: invalidate fingerprint after each authentication
-    </releaseNotes>
-    <dependencies>
-      <group targetFramework="portable-net45+win8+wpa81+wp8">
-        <dependency id="MvvmCross.Platform" version="4.0.0" />
-      </group>
-      <group targetFramework="MonoAndroid">
-        <dependency id="MvvmCross.Platform" version="4.0.0" />
-      </group>
-      <group targetFramework="Xamarin.iOS">
-        <dependency id="MvvmCross.Platform" version="4.0.0" />
-      </group>
-      <group targetFramework="UAP10">
-        <dependency id="MvvmCross.Platform" version="4.0.0" />
-      </group>
-    </dependencies>
-  </metadata>
-  <files>
-    <!-- PCL -->
-    <file src="mvx\pcl\MvvmCross.Plugins.Fingerprint.*" target="lib\portable-net45+win8+wpa81+wp8" />
-    <file src="pcl\Plugin.Fingerprint.Abstractions.*" target="lib\portable-net45+win8+wpa81+wp8" />
-    <!-- Android -->
-    <file src="mvx\android\MvvmCross.Plugins.Fingerprint.*" target="lib\MonoAndroid" />
-    <file src="mvx\android\Plugin.Fingerprint.*" target="lib\MonoAndroid" />
-    <file src="mvx\android\FingerprintPluginBootstrap.cs.pp" target="content\MonoAndroid\Bootstrap\FingerprintPluginBootstrap.cs.pp" />
-    <!-- iOS -->
-    <file src="mvx\ios\MvvmCross.Plugins.Fingerprint.*" target="lib\Xamarin.iOS10" />
-    <file src="mvx\ios\Plugin.Fingerprint.Abstractions.*" target="lib\Xamarin.iOS10" />
-    <file src="mvx\ios\Plugin.Fingerprint.*" target="lib\Xamarin.iOS10" />
-    <file src="mvx\ios\FingerprintPluginBootstrap.cs.pp" target="content\Xamarin.iOS10\Bootstrap\FingerprintPluginBootstrap.cs.pp" />
-    <!-- UWP -->
-    <file src="mvx\uwp\MvvmCross.Plugins.Fingerprint.*" target="lib\UAP10" />
-    <file src="mvx\uwp\Plugin.Fingerprint.*" target="lib\UAP10" />
-    <file src="mvx\uwp\Plugin.Fingerprint\Properties\*.rd.xml" target="lib\UAP10\Plugin.Fingerprint\Properties" />
-    <file src="mvx\uwp\MvvmCross.Plugins.Fingerprint.WindowsUWP\Properties\*.rd.xml" target="lib\UAP10\MvvmCross.Plugins.Fingerprint.WindowsUWP\Properties" />
-    <file src="mvx\uwp\FingerprintPluginBootstrap.cs.pp" target="content\UAP10\Bootstrap\FingerprintPluginBootstrap.cs.pp" />
-  </files>
-</package>
