--- src/.build/Plugin.Fingerprint.nuspec
@@ -1,84 +1,0 @@
-<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-  <metadata>
-    <id>Plugin.Fingerprint</id>
-    <version>1.4.7</version>
-    <title>Fingerprint Plugin for Xamarin</title>
-    <authors>Sven-Michael Stübe</authors>
-    <owners>Sven-Michael Stübe</owners>
-    <licenseUrl>https://github.com/smstuebe/xamarin-fingerprint/blob/master/LICENSE</licenseUrl>
-    <projectUrl>https://github.com/smstuebe/xamarin-fingerprint</projectUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Authenticate a user via fingerprint from a cross platform API.</description>
-    <summary>Authenticate a user via fingerprint from a cross platform API.</summary>
-    <tags>xamarin mvvmcross mvvm monodroid Xamarin.iOS10 mac Xamarin.Mac ios fingerprint authenticate Xamarin.Android android touchid biometric windows uwp</tags>
-    <iconUrl>https://raw.githubusercontent.com/smstuebe/xamarin-fingerprint/master/doc/xamarin_fingerprint.png</iconUrl>
-    <releaseNotes>
-      [1.4.7]
-      - #101 FingerprintAuthenticationResultStatus and FingerprintAvailability now can express that the user has denied the access to the authentication method
-      [1.4.6]
-      - CrossFingerprint is now setable for mocking during unit tests (thanks @ArtjomP)
-      - #51 allows setting the `DefaultColor` to change the color of the fingerprint icon
-      - #86 android fallback button is now hidden if AllowAlternativeAuthentication is false (thanks @fedemkr)
-      - #91 new API `GetAuthenticationTypeAsync` to retrieve the biometric auth type
-      - #84 missing resource ids should be available again
-      - #83 prevent NullRefereneceException and catch others if service is not available on Android
-      - #75 fix crash at onPause on Samsung devices
-      - #73 display help string for recoverable errors
-      - #29 fix crash on device lock during authentication
-      [1.4.5]
-      - #53, #70 fixed possible crash on some samsung devices without fingerprint sensor
-      [1.4.4]
-      - #60 allow alternative authentication via PIN / password (iOS only)
-      - allow custon fallback title on Mac
-      [1.4.3]
-      - #45: fixed UWP nuget package
-      [1.4.2]
-      - #43: fixed crash when fast tapping on cancel or fallback on Android
-      [1.4.1]
-      - #40: fixed crash on iOS &lt; 8.0
-      [1.4.0]
-      - Xamarin.Mac support
-      [1.3.0]
-      - support for Samsung devices
-      - support for .NET standard
-      - support for dialog localization
-      - animations on Android
-      - bugfixes
-      [1.2.0]
-      - iOS: invalidate fingerprint after each authentication
-    </releaseNotes>
-    <dependencies>
-      <group targetFramework="netstandard1.0">
-        <dependency id="NETStandard.Library" version="1.6.0" />
-      </group>
-      <group targetFramework="portable-net45+win8+wpa81+wp8">
-      </group>
-      <group targetFramework="Xamarin.Mac20">
-      </group>
-    </dependencies>
-  </metadata>
-  <files>
-    <!-- PCL / net standard -->
-    <file src="pcl\Plugin.Fingerprint.*" target="lib\portable-net45+win8+wpa81+wp8" />
-    <file src="pcl\Plugin.Fingerprint.*" target="lib\netstandard1.0" />
-    <!-- .NET 4.5 -->
-    <file src="null\Plugin.Fingerprint.*" target="lib\net45" />
-    <!-- Android -->
-    <file src="android\Plugin.Fingerprint.*" target="lib\MonoAndroid" />
-    <!-- iOS -->
-    <file src="ios\Plugin.Fingerprint.*" target="lib\Xamarin.iOS10" />
-    <!-- Mac -->
-    <file src="mac\Plugin.Fingerprint.*" target="lib\Xamarin.Mac20" />
-    <!-- UWP -->
-    <file src="uwp\Plugin.Fingerprint.*" target="lib\UAP10" />
-    <file src="uwp\Plugin.Fingerprint\Properties\*.rd.xml" target="lib\UAP10\Plugin.Fingerprint\Properties" />
-    <!-- Win 8 Store -->
-    <file src="null\Plugin.Fingerprint.*" target="lib\netcore45" />
-    <!-- Win 8.1 -->
-    <file src="null\Plugin.Fingerprint.*" target="lib\win81" />
-    <!-- Win Phone 8 -->
-    <file src="null\Plugin.Fingerprint.*" target="lib\wp8" />
-    <!-- Win Phone 8.1 -->
-    <file src="null\Plugin.Fingerprint.*" target="lib\wpa81" />
-  </files>
-</package>
