+++ src/Fingerprint.sln
@@ -4,6 +4,10 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 15.0.27130.2027
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{3015FCD9-4308-47F9-BC7B-1C060FCA31A9}"
+	ProjectSection(SolutionItems) = preProject
+		Sample\Directory.build.props = Sample\Directory.build.props
+		Sample\Directory.build.targets = Sample\Directory.build.targets
+	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SMS.Fingerprint.Sample.Droid", "Sample\SMS.Fingerprint.Sample.Droid\SMS.Fingerprint.Sample.Droid.csproj", "{7096490F-E271-412B-8376-D58C226AB9FE}"
 EndProject
@@ -11,28 +15,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SMS.Fingerprint.Sample.iOS"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SMS.Fingerprint.Sample", "Sample\SMS.Fingerprint.Sample\SMS.Fingerprint.Sample.csproj", "{02094FA7-945E-43BB-AEB9-203C2D018E86}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".build", ".build", "{44D4F832-D019-4500-9AEE-BA7038FCBD91}"
-	ProjectSection(SolutionItems) = preProject
-		.build\build.bat = .build\build.bat
-		.build\build.fsx = .build\build.fsx
-		.build\MvvmCross.Plugins.Fingerprint.nuspec = .build\MvvmCross.Plugins.Fingerprint.nuspec
-		.build\Plugin.Fingerprint.nuspec = .build\Plugin.Fingerprint.nuspec
-	EndProjectSection
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{23667AB3-6148-4B1C-8DE1-EF92273D25E4}"
-	ProjectSection(SolutionItems) = preProject
-		.nuget\NuGet.Config = .nuget\NuGet.Config
-		.nuget\NuGet.exe = .nuget\NuGet.exe
-		.nuget\NuGet.targets = .nuget\NuGet.targets
-	EndProjectSection
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "MvvmCross", "MvvmCross", "{BA5BB115-619A-4189-8F85-E8D6E5100929}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SMS.Fingerprint.Sample.UWP", "Sample\SMS.Fingerprint.Sample.UWP\SMS.Fingerprint.Sample.UWP.csproj", "{25F504C9-3250-4FD7-A6AA-A1FC11F24C08}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MvvmCross.Plugins.Fingerprint", "MvvmCross.Plugins.Fingerprint\MvvmCross.Plugins.Fingerprint.csproj", "{99AF4877-02DF-4DAE-A524-4056BAE82614}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Plugin.Fingerprint.Android.Samsung", "Plugin.Fingerprint.Android.Samsung\Plugin.Fingerprint.Android.Samsung.csproj", "{E81F50D5-95C3-4D5A-9397-F79946912B5D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Plugin.Fingerprint.Android.Samsung", "Plugin.Fingerprint.Android.Samsung\Plugin.Fingerprint.Android.Samsung.csproj", "{E81F50D5-95C3-4D5A-9397-F79946912B5D}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SMS.Fingerprint.Sample.Mac", "Sample\SMS.Fingerprint.Sample.Mac\SMS.Fingerprint.Sample.Mac.csproj", "{F8C7D7D8-7409-4D03-BAC1-776E91406CDE}"
 EndProject
@@ -43,6 +30,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".doc", ".doc", "{7674D66B-F
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Plugin.Fingerprint", "Plugin.Fingerprint\Plugin.Fingerprint.csproj", "{E519A552-0A87-4E3C-965A-721426E32637}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SolutionItems", "SolutionItems", "{9916AC77-31D2-4F3E-BF65-A78BA21C3CD6}"
+	ProjectSection(SolutionItems) = preProject
+		Directory.build.props = Directory.build.props
+		Directory.build.targets = Directory.build.targets
+		global.json = global.json
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -487,7 +481,6 @@ Global
 		{DA111357-2B43-4A94-9478-B42A2BF8DE7E} = {3015FCD9-4308-47F9-BC7B-1C060FCA31A9}
 		{02094FA7-945E-43BB-AEB9-203C2D018E86} = {3015FCD9-4308-47F9-BC7B-1C060FCA31A9}
 		{25F504C9-3250-4FD7-A6AA-A1FC11F24C08} = {3015FCD9-4308-47F9-BC7B-1C060FCA31A9}
-		{99AF4877-02DF-4DAE-A524-4056BAE82614} = {BA5BB115-619A-4189-8F85-E8D6E5100929}
 		{F8C7D7D8-7409-4D03-BAC1-776E91406CDE} = {3015FCD9-4308-47F9-BC7B-1C060FCA31A9}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
