--- src/MvvmCross.Plugins.Fingerprint/Platforms/UWP/Plugin.cs
@@ -1,15 +1,0 @@
-﻿using MvvmCross.Plugin;
-using Plugin.Fingerprint;
-
-namespace MvvmCross.Plugins.Fingerprint.WindowsUWP
-{
-    [MvxPlugin]
-    [Preserve(AllMembers = true)]
-    public class Plugin : IMvxPlugin
-    {
-        public void Load()
-        {
-            Mvx.LazyConstructAndRegisterSingleton(() => CrossFingerprint.Current);
-        }
-    }
-}
