+++ src/MvvmCross.Plugins.Fingerprint/Platforms/Uap/Plugin.cs
@@ -1,0 +1,15 @@
+﻿using MvvmCross.Plugin;
+using Plugin.Fingerprint;
+
+namespace MvvmCross.Plugins.Fingerprint.WindowsUWP
+{
+    [MvxPlugin]
+    [Preserve(AllMembers = true)]
+    public class Plugin : IMvxPlugin
+    {
+        public void Load()
+        {
+            Mvx.LazyConstructAndRegisterSingleton(() => CrossFingerprint.Current);
+        }
+    }
+}
